+++ app/controllers/carto/api/mapcaps_controller.rb
@@ -58,7 +58,7 @@ module Carto
       end
 
       def owners_only
-        raise Carto::UnauthorizedError.new unless @visualization.is_writable_by_user(current_user)
+        raise Carto::UnauthorizedError.new unless @visualization.writable_by?(current_user)
       end
 
       MAX_MAPCAPS_PER_MAP = 1
@@ -78,7 +78,8 @@ module Carto
       end
 
       def track_published_map
-        Carto::Tracking::Events::PublishedMap.new(current_user, @visualization).report
+        Carto::Tracking::Events::PublishedMap.new(user_id: current_user.id,
+                                                  visualization_id: @visualization.id).report
       end
     end
   end