+++ app/controllers/carto/builder/datasets_controller.rb
@@ -44,7 +44,7 @@ module Carto
       end
 
       def authors_only
-        unauthorized unless current_user && @canonical_visualization.is_writable_by_user(current_user)
+        unauthorized unless current_user && @canonical_visualization.writable_by?(current_user)
       end
 
       def load_user_table
@@ -61,7 +61,7 @@ module Carto
       end
 
       def track_dataset_visit
-        Carto::Tracking::Events::VisitedPrivateDataset.new(current_user).report
+        Carto::Tracking::Events::VisitedPrivatePage.new(user_id: current_user.id).report
       end
     end
   end