+++ app/controllers/sessions_controller.rb
@@ -1,6 +1,7 @@
 # encoding: UTF-8
 require_dependency 'google_plus_config'
 require_dependency 'google_plus_api'
+require_dependency 'oauth/github/config'
 
 require_relative '../../lib/user_account_creator'
 require_relative '../../lib/cartodb/stats/authentication'
@@ -17,7 +18,8 @@ class SessionsController < ApplicationController
                      only: [:account_token_authentication_error, :ldap_user_not_at_cartodb]
 
   before_filter :load_organization
-  before_filter :initialize_google_plus_config
+  before_filter :initialize_google_plus_config,
+                :initialize_github_config
   before_filter :api_authorization_required, only: :show
 
   def new
@@ -133,6 +135,15 @@ class SessionsController < ApplicationController
 
     button_color = @organization.nil? || @organization.color.nil? ? nil : organization_color(@organization)
     @google_plus_config = ::GooglePlusConfig.instance(CartoDB, Cartodb.config, signup_action, 'google_access_token', button_color)
+  end
+
+  def initialize_github_config
+    unless @organization && !@organization.auth_github_enabled
+      @github_config = Carto::Github::Config.instance(form_authenticity_token,
+                                                      invitation_token: params[:invitation_token],
+                                                      organization_name: @organization.try(:name))
+      @button_color = @organization && @organization.color ? organization_color(@organization) : nil
+    end
   end
 
   def extract_username(request, params)